CCZ4: Reinstate Constraints.impl.hpp
Because SYCL requires all device functions to be defined within the
same translation unit or qualified with SYCL_EXTERNAL, I have
moved Constraints.cpp back to Constraints.impl.hpp.

I also added back the AMREX_FORCE_INLINE qualifiers.
