+++ Source/CCZ4/Make.package
@@ -1,5 +1,3 @@
-GRTECLYN_CEXE_sources += Constraints.cpp
-
 GRTECLYN_CEXE_headers += ADMConformalVars.hpp \
                          BSSNVars.hpp \
                          CCZ4Geometry.hpp \
